pyproject.toml
[tool.poetry]
name = "glinthub"
version = "0.8.0"
description = "API for an online shop selling gold products"
authors = ["Ali Sarbakhshi"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.3"
celery = "^5.4.0"
django-filter = "^24.3"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
requests = "^2.32.3"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^11.0.0"
python-dotenv = "^1.0.1"
redis = "^5.2.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
faker = "^33.1.0"
pytest-django = "^4.9.0"

[tool.poetry.group.production.dependencies]
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.development.dependencies]
django-debug-toolbar = "^4.4.6"
pre-commit = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Pytest configurations
[tool.pytest.ini_options]
python_files = 'tests.py'
required_plugins = 'pytest-django'
DJANGO_SETTINGS_MODULE = 'config.settings.development' # pytest-django required settings
#TODO: Tried converage using pytest-cov and it doesn't play well with pytest-django. Check later using the coverage library
# or alternative libraries
